fix warnings for printf-style calls, some of which are real fixes for 64-bit mode.

+++ src/hack.hxx
@@ -20,7 +20,7 @@ static void canvas_reflecttitle(t_canvas *x)
     }
     else namebuf[0] = 0;
     sys_vgui("wm title .x%lx {%s%c%s - %s}\n",
-        x, x->gl_name->s_name, (x->gl_dirty? '*' : ' '), namebuf,
+        long(x), x->gl_name->s_name, (x->gl_dirty? '*' : ' '), namebuf,
             canvas_getdir(x)->s_name);
 }
 #else
